* nasl/nasl_smb.c (nasl_smb_connect): Don't cast handle to int.
* nasl/nasl_var.c (var2int, get_int_var_by_num, get_int_var_by_name)
(get_int_local_var_by_name): Handle long integers.

* nasl/nasl_lex_ctxt.h: Adjust functions headers.

* nasl/openvas_smb_interface.h (SMB_HANDLE): Define as long int.

git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-libraries@27140 423fd1db-d629-0410-8442-d21db03e70f4

+++ nasl/nasl_smb.c
@@ -136,7 +136,7 @@ nasl_smb_connect (lex_ctxt * lexic)
       return NULL;
     }
 
-  retc->x.i_val = (int) handle;
+  retc->x.i_val = handle;
   return retc;
 }
